Fix deletion

+++ app/Models/User.php
@@ -169,19 +169,27 @@ class User extends Authenticatable implements WebAuthnAuthenticatable
 	 */
 	public function delete(): bool
 	{
-		$now = Carbon::now();
-		$newOwnerID = Auth::id() ?? throw new UnauthenticatedException();
-
 		/** @var HasMany[] $ownershipRelations */
 		$ownershipRelations = [$this->photos(), $this->albums()];
+		$hasAny = false;
 
 		foreach ($ownershipRelations as $relation) {
-			// We must also update the `updated_at` column of the related
-			// models in case clients have cached these models.
-			$relation->update([
-				$relation->getForeignKeyName() => $newOwnerID,
-				$relation->getRelated()->getUpdatedAtColumn() => $relation->getRelated()->fromDateTime($now),
-			]);
+			$hasAny = $hasAny || $relation->count() > 0;
+		}
+
+		if ($hasAny) {
+			// only try update relations if there are any to allow deleting users from migrations (relations are moved before deleting)
+			$now = Carbon::now();
+			$newOwnerID = Auth::id() ?? throw new UnauthenticatedException();
+
+			foreach ($ownershipRelations as $relation) {
+				// We must also update the `updated_at` column of the related
+				// models in case clients have cached these models.
+				$relation->update([
+					$relation->getForeignKeyName() => $newOwnerID,
+					$relation->getRelated()->getUpdatedAtColumn() => $relation->getRelated()->fromDateTime($now),
+				]);
+			}
 		}
 
 		$this->shared()->delete();
